requirements.txt
playwright
python-dotenv
coloredlogs
tk
requests
groq
open-interpreter
customtkinter
pillow
xlib
pyobjc
